backend/pyproject.toml
[project]
name = "compliance-workflow-demo"
version = "1.9.0"
description = "Compliance DSL + outage-resilient LLM router demo ([Client] interview)"
requires-python = ">=3.12,<3.13"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "pydantic>=2.9",
  "httpx>=0.27",
  "tenacity>=9",
  "psycopg[binary]>=3.2",
  "pypdf>=5",
  "tiktoken>=0.8",
  "reportlab>=4.2",
  "opentelemetry-sdk>=1.28",
  "opentelemetry-exporter-otlp-proto-grpc>=1.28",
  "opentelemetry-instrumentation-fastapi>=0.49b0",
  "pyyaml>=6",
  "python-dotenv>=1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/compliance_workflow_demo"]

[dependency-groups]
dev = [
  "ruff>=0.6",
  "pytest>=8",
  "pytest-asyncio>=0.24",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
